finished tasks

+++ controllers/contactsControllers.js
@@ -3,7 +3,17 @@ import HttpError from "../helpers/HttpError.js";
 
 export const getAllContacts = async (req, res, next) => {
   try {
-    const result = await contactsService.listContacts();
+    const { page = 1, limit = 20, favorite = true } = req.query;
+    const { _id: owner } = req.user;
+    const filter = { owner, favorite };
+    const fields = "-createdAt -updatedAt";
+    const skip = (page - 1) * limit;
+    const settings = { skip, limit };
+    const result = await contactsService.listContacts({
+      filter,
+      fields,
+      settings,
+    });
     res.json(result);
   } catch (error) {
     next(error);
@@ -13,7 +23,8 @@ export const getAllContacts = async (req, res, next) => {
 export const getOneContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await contactsService.getContactById(id);
+    const { _id: owner } = req.user;
+    const result = await contactsService.getContactById({ _id: id, owner });
     if (!result) {
       throw HttpError(404, "Not found");
     }
@@ -26,7 +37,8 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await contactsService.removeContact(id);
+    const { _id: owner } = req.user;
+    const result = await contactsService.removeContact({ _id: id, owner });
     if (!result) {
       throw HttpError(404, "Not found");
     }
@@ -38,7 +50,8 @@ export const deleteContact = async (req, res, next) => {
 
 export const createContact = async (req, res, next) => {
   try {
-    const result = await contactsService.addContact(req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.addContact({ ...req.body, owner });
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -48,7 +61,11 @@ export const createContact = async (req, res, next) => {
 export const updateContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await contactsService.updateContact(id, req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.updateContact(
+      { _id: id, owner },
+      req.body
+    );
     if (!result) {
       throw HttpError(404, "Not found");
     }
@@ -61,7 +78,11 @@ export const updateContact = async (req, res, next) => {
 export const updateStatusContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await contactsService.updateContact(id, req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.updateContact(
+      { _id: id, owner },
+      req.body
+    );
 
     if (!result) {
       throw HttpError(404, "Not found");
